* commit '25b4f4b445de867fab15c6f353e8a616d6ebe2c9':
  JA-391 #comment
  JA-402 #comment
  JA-402 #comment
  JA-391 #comment
  JA-391 #comment

+++ app/src/main/java/life/plank/juna/zone/util/UIDisplayUtil.java
@@ -303,7 +303,7 @@ public class UIDisplayUtil {
                 true);
     }
 
-    public static float getDp(Context context, int pixels) {
+    public static float getDp(Context context, float pixels) {
         return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, pixels, context.getResources().getDisplayMetrics());
     }
